requirements.txt
GitPython==3.1.45
Jinja2==3.1.6
MarkupSafe==3.0.3
Pint==0.25
PyMuPDF==1.26.5
Pygments==2.19.2
RapidFuzz==3.14.1
Unidecode==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
flexcache==0.3
flexparser==0.4
gitdb==4.0.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.11.1
jsonschema-specifications==2025.9.1
jsonschema==4.25.1
lxml==6.0.2
narwhals==2.10.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.1
pluggy==1.6.0
protobuf==6.33.0
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pypdfium2==5.0.0
pytesseract==0.3.13
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
qrcode==8.2
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rpds-py==0.28.0
segno==1.6.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
xlrd==2.0.2
uvicorn
fastapi